+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `!adminlist` now shows the last time the bot has seen each admin and on which alt.
 - Add new option `--strict` to make SQLite checks more strict. This is mainly for development purpose.
 - Add new prometheus metric `states{type="raid_lock"}`
+- The `LOOT_MODULE` now keeps a full history of what was rolled when, and who won what on which roll. You can search this history by using `!loot history`, `!loot history <number>|last`, `!loot search winner=Nady` and `!loot search item=leg`.
 
 ### Changed